pyproject.toml
[tool.poetry]
name = "pyconlyse"
version = "0.2.0"
description = "This is PYCONLYSE tango based application for controlling ELYSE experiment."
authors = ["saldenisov <saldenisov@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = "^3.7"
configobj = "^5"
colorcet = "^3"
gpiozero = "^2"
pigpio = "^1.78"
guiqwt = "^4.4"
ifaddr = "^0.2"
imutils = "^0.5"
imageio = "^2.33"
msgpack = "^1.0"
msl-loadlib = "^0.10"
lxml = "^4.9"
keyboard = "^0.13"
opencv-python = "^4.8"
pandas = "^2.0"
pyqtgraph = "^0.13"
PyQt5 = "<5.13"
PyQt5-sip = "^12"
pyepics = "^3.5"
pypylon = "^3"
pyserial = "^3.5"
PyTango = "^9.4.1"
requests = "^2.3"
scipy = "^1.8"
taurus = "^5.1"
taurus_pyqtgraph = "^0.7"
spyder = "^5.1"
simple-pid = "^2.0"
vitables = "^3.0"
QtPy = ">=1.3"
PythonQwt = ">=0.5"
guidata = ">=1.7"
Pillow = "^10"
pyarrow = "^14.0"
h5py = "^3"






[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
